% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{akan}
\alias{akan}
\title{Data from Akan et al. (2018)}
\format{
a data frame with 270 rows and 3 variables:
\describe{
  \item{subject}{subject id}
  \item{condition}{experimental condition (itemtest, restudy, control)}
  \item{recalled}{dependent variable}
}
}
\source{
Akan, M., Stanley, S. E., & Benjamin, A. S. (2018). Testing enhances
  memory for context. Journal of Memory and Language, 103, 19–27.
  \url{https://doi.org/10.1016/j.jml.2018.07.003}
}
\usage{
akan
}
\description{
For the entire dataset and analysis scripts see: \url{https://osf.io/bqr5f/}
}
\keyword{datasets}
