% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_contrast.R
\name{calc_r_alerting}
\alias{calc_r_alerting}
\title{Calculate r_alerting from r_contrast and r_effectsize}
\usage{
calc_r_alerting(r_contrast, r_effectsize)
}
\arguments{
\item{r_contrast}{what it says}

\item{r_effectsize}{what it says}
}
\description{
Convenience function to transform effect sizes in contrast analyses.
}
