% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_contrast.R
\name{calc_r_alerting_from_f}
\alias{calc_r_alerting_from_f}
\title{Calculate r_alerting from F-values}
\usage{
calc_r_alerting_from_f(f_contrast, f_between, df_between)
}
\arguments{
\item{f_contrast}{F value from contrast analysis}

\item{f_between}{F value from ANOVA (one between variable!)}

\item{df_between}{degrees of freedom of ANOVA}
}
\description{
Convenience function to calculate effect sizes in contrast analyses.
}
