% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_contrast.R
\name{calc_r_contrast}
\alias{calc_r_contrast}
\title{Calculate r_contrast from r_alerting and r_effectsize}
\usage{
calc_r_contrast(r_alerting, r_effectsize)
}
\arguments{
\item{r_alerting}{what it says}

\item{r_effectsize}{what it says}
}
\description{
Convenience function to transform effect sizes in contrast analyses.
}
