% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_contrast.R
\name{calc_r_effectsize}
\alias{calc_r_effectsize}
\title{Calculate r_effectsize from r_contrast and r_alerting}
\usage{
calc_r_effectsize(r_alerting, r_contrast)
}
\arguments{
\item{r_alerting}{what it says}

\item{r_contrast}{what it says}
}
\description{
Convenience function to transform effect sizes in contrast analyses.
}
