% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cofad-package.R
\docType{package}
\name{cofad-package}
\alias{cofad}
\alias{cofad-package}
\title{cofad: Contrast Analyses for Factorial Designs}
\description{
Contrast analysis for factorial designs provides an alternative to the traditional ANOVA approach, offering the distinct advantage of testing targeted hypotheses. The foundation of this package is primarily rooted in the works of Rosenthal, Rosnow, and Rubin (2000, ISBN: 978-0521659802) as well as Sedlmeier and Renkewitz (2018, ISBN: 978-3868943214).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/johannes-titz/cofad}
}

}
\author{
\strong{Maintainer}: Johannes Titz \email{johannes.titz@gmail.com}

Authors:
\itemize{
  \item Johannes Titz \email{johannes.titz@gmail.com}
  \item Markus Burkhardt \email{markus.burkhardt@psychologie.tu-chemnitz.de}
}

Other contributors:
\itemize{
  \item Mirka Henninger \email{mail@mirka-henninger.de} [contributor]
  \item Simone Malejka [contributor]
}

}
\keyword{internal}
