% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{furr_p4}
\alias{furr_p4}
\title{Empathy data set by Furr (2004)}
\format{
a data frame with 20 rows and 2 columns
\describe{
  \item{empathy}{Empathy rating}
  \item{major}{major of student}
}
}
\source{
Furr, R. M. (2004). Interpreting effect sizes in contrast analysis.
  Understanding Statistics, 3, 1–25.
  https://doi.org/10.1207/s15328031us0301_1
}
\usage{
furr_p4
}
\description{
fictitious data set on empathy ratings of students from different majors
}
\keyword{datasets}
