% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{haans_within1by4}
\alias{haans_within1by4}
\title{Haans within data example}
\format{
a data frame with 20 rows and 3 variables:
\describe{
  \item{person}{person id}
  \item{name}{group name (sitting row 1 to 4)}
  \item{value}{dv, final exam grade}
}
}
\usage{
haans_within1by4
}
\description{
Fictitious data set from Haans, A. (2018). Contrast Analysis: A Tutorial.
https://doi.org/10.7275/7DEY-ZD62
}
\keyword{datasets}
