% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.cofad_bw}
\alias{print.cofad_bw}
\title{Output of between subject design contrast analysis}
\usage{
\method{print}{cofad_bw}(x, ...)
}
\arguments{
\item{x}{output of calc_contrast}

\item{...}{further arguments}
}
\value{
Displays the significance of the contrast analysis.
given.
}
\description{
Output of between subject design contrast analysis
}
