% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.cofad_wi}
\alias{print.cofad_wi}
\title{Output of a within subject design contrast analysis}
\usage{
\method{print}{cofad_wi}(x, ...)
}
\arguments{
\item{x}{output of calc_contrast}

\item{...}{further arguments}
}
\value{
Displays the significance of the contrast analysis. The
contrast weights, the corresponding group, and an effect size are given.
}
\description{
Output of a within subject design contrast analysis
}
