% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rosenthal_chap5_q2}
\alias{rosenthal_chap5_q2}
\title{Complexity data set by Rosenthal and Rosnow (2000)}
\format{
a data frame with 12 rows and 4 columns
\describe{
  \item{dv}{dependent variable: rating of degree of complexity of social
  interaction from a series of clips}
  \item{id}{unique identifier of participant}
  \item{within}{within variable: complexity of interaction (low, medium
  high)}
  \item{between}{between variable: cognitive complexity of participant (high
  or low)}
}
}
\source{
Rosenthal, R., Rosnow, R. L., & Rubin, D. B. (2000). Contrasts and
  Effect Sizes in Behavioral Research: A Correlational Approach. Cambridge
  University Press.
}
\usage{
data(rosenthal_chap5_q2)
}
\description{
Exercise 2 from Chapter 5 (table on p. 147) in Rosenthal and Rosnow (2000)
}
\keyword{datasets}
