% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rosenthal_tbl31}
\alias{rosenthal_tbl31}
\title{Data set by Rosenthal and Rosnow (2000)}
\format{
a data frame with 20 rows and 2 columns
\describe{
  \item{dv}{dependent variable}
  \item{between}{group (A, B, C, D))}
}
}
\source{
Rosenthal, R., Rosnow, R. L., & Rubin, D. B. (2000). Contrasts and
  Effect Sizes in Behavioral Research: A Correlational Approach. Cambridge
  University Press.
}
\usage{
rosenthal_tbl31
}
\description{
Table 3.1 in Rosenthal and Rosnow (2000) on p. 38.
}
\keyword{datasets}
