% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rosenthal_tbl59}
\alias{rosenthal_tbl59}
\title{Therapy data set by Rosenthal and Rosnow (2000)}
\format{
a data frame with 12 rows and 4 columns
\describe{
  \item{id}{unique identifier}
  \item{dv}{dependent variable}
  \item{med}{within variable: medication (treatment or placebo)}
  \item{pt}{between variable: psychotherapy (treatment or placebo)}
}
}
\source{
Rosenthal, R., Rosnow, R. L., & Rubin, D. B. (2000). Contrasts and
  Effect Sizes in Behavioral Research: A Correlational Approach. Cambridge
  University Press.
}
\usage{
data(rosenthal_tbl59)
}
\description{
Table 5.9 in Rosenthal and Rosnow (2000)
}
\keyword{datasets}
