% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rosenthal_tbl68}
\alias{rosenthal_tbl68}
\title{Data set by Rosenthal and Rosnow (2000)}
\format{
a data frame with 32 rows and 4 columns
\describe{
  \item{id}{unique identifier of participant}
  \item{dv}{dependent variable}
  \item{within}{within variable}
  \item{between}{between variable}
}
}
\source{
Rosenthal, R., Rosnow, R. L., & Rubin, D. B. (2000). Contrasts and
  Effect Sizes in Behavioral Research: A Correlational Approach. Cambridge
  University Press.
}
\usage{
data(rosenthal_tbl68)
}
\description{
Fictitious example of children ability, Table 6.8 in Rosenthal and Rosnow
(2000)
}
\keyword{datasets}
