% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schwoebel}
\alias{schwoebel}
\title{Data from Schwoebel et al. (2018)}
\format{
a data frame with 64 rows and 2 variables:
\describe{
  \item{condition}{experimental condition (massed-same, massed-different,
    spaced-same, spaced-different)}
  \item{percent_recalled}{dependent variable}
}
}
\source{
Schwoebel, J., Depperman, A. K., & Scott, J. L. (2018). Distinct
  episodic contexts enhance retrieval-based learning. Memory, 26(9),
  1291–1296. \url{https://doi.org/10.1080/09658211.2018.1464190}
}
\usage{
schwoebel
}
\description{
For the entire dataset and analysis scripts see:
}
\keyword{datasets}
