% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sedlmeier_p525}
\alias{sedlmeier_p525}
\title{Problem solving data set by Sedlmeier & Renkewitz (2018)}
\format{
a data frame with 15 rows and 3 columns
\describe{
  \item{lsg}{dv, number of solved exercises}
  \item{between}{group, KT=no training, JT=boys-specific, MT=girls-specific}
  \item{lambda}{lambdas used for this example}
}
}
\source{
Sedlmeier, P., & Renkewitz, F. (2018). Forschungsmethoden und
  Statistik für Psychologen und Sozialwissenschaftler (3rd ed.). Pearson
  Studium.
}
\usage{
data(sedlmeier_p525)
}
\description{
Example 16.2, Table 16.1 in Sedlmeier & Renkewitz (2018). Fictitious data set
with 15 boys divided into three groups (no training, boys-specific material,
girls-specific training material). The DV is the number of solved problem
tasks. Table 16.2 in the book repeats these observations and adds the three
sets of contrast weights discussed in the example.
}
\keyword{datasets}
