% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sedlmeier_p537}
\alias{sedlmeier_p537}
\title{Music data set by Sedlmeier & Renkewitz (2018)}
\format{
a data frame with 32 rows and 3 columns
\describe{
  \item{reading_test}{dependent variable}
  \item{participant}{unique id}
  \item{music}{within variable}
}
}
\source{
Sedlmeier, P., & Renkewitz, F. (2018). Forschungsmethoden und
  Statistik für Psychologen und Sozialwissenschaftler (3rd ed.). Pearson
  Studium.
}
\usage{
sedlmeier_p537
}
\description{
Example 16.6, table 16.5 in Sedlmeier & Renkewitz (2018). Fictitious data set
with 8 participants that listened to no music, white noise, classical music,
and jazz music (within). The DV is a reading test.
}
\keyword{datasets}
