% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary.cofad_bw}
\alias{summary.cofad_bw}
\title{Summary of between subject design contrast analysis}
\usage{
\method{summary}{cofad_bw}(object, ...)
}
\arguments{
\item{object}{output of calc_contrast}

\item{...}{further arguments}
}
\value{
Displays ANOVA table of the contrastanalysis
and the typical effectsizes.
}
\description{
Summary of between subject design contrast analysis
}
