% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{akan}
\alias{akan}
\title{Data from Akan et al. (2018), experiment 2B}
\format{
a data frame with 270 rows and 3 variables:
\describe{
  \item{subject}{subject id}
  \item{condition}{experimental condition (test, restudy, control)}
  \item{recalled}{dependent variable}
}
}
\source{
Akan, M., Stanley, S. E., & Benjamin, A. S. (2018). Testing enhances
  memory for context. Journal of Memory and Language, 103, 19–27.
  \url{https://doi.org/10.1016/j.jml.2018.07.003}
}
\usage{
akan
}
\description{
Data contains information from a within-subjects experiment with N = 90
participants. The goal of the experiment was to investigate the benefits of
retrieval practice on memory performance. For the entire dataset and analysis
scripts see: \url{https://osf.io/bqr5f/}. The data was licsensed under CC-BY
4.0 Melisa Akan, Aaron Benjamin.
}
\keyword{datasets}
