% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_contrast.R
\name{calc_contrast}
\alias{calc_contrast}
\title{Calculate contrast analysis for factorial designs}
\usage{
calc_contrast(variable, between = NULL, between_levels = NULL,
  lambda_between = NULL, within = NULL, within_levels = NULL,
  lambda_within = NULL, ID = NULL, data = NULL)
}
\arguments{
\item{variable}{The dependent variable. This must be numeric.}

\item{between}{The independent variable which cut yout data into
independent groups. This must be a factor.}

\item{between_levels}{The names of the between groups. The
sqequence of between_levels must correspondent to the sequence
of  between_lambda. This must be a character}

\item{lambda_between}{The Contrastweights. This must be numeric and
correspondent to the sequence of between_levels. If
lambda_between not sum up to zero, this will be done automatically}

\item{within}{The independent variable which cut yout data into
dependent groups. This must be a factor.}

\item{within_levels}{The names of the within groups. The
sqequence of within_levels must correspondent to the sequence
of  within_lambda. This must be a character}

\item{lambda_within}{The Contrastweights. This must be numeric and
correspondent to the sequence of within_levels. If
lambda_within not sum up to zero, this will be done automatically}

\item{ID}{The cases / subjects connectet to within.}

\item{data}{Optional argument for a data.frame containing variable
and group.}
}
\value{
Calculates the significance of the contrast analysis.
given to evaluate the analysis.
}
\description{
Calculate contrast analysis for factorial designs
}
\details{
For more than one wihtin or between variable,
contrastweights have to be connected See example 4 or 5
correspondent to the expected multifactorial effect.
}
\examples{
# Exampel for between-subjects-design Table 3.1 from
# Rosenthal, Rosnow and Rubin (2001)

tab31 <- data.frame(
  Val = c(2, 6, 8, 4,10, 6, 8, 10, 4, 12, 8,
    16, 10, 14, 12, 12,  18, 14, 20, 16),
  Let = rep(c("A", "B", "C", "D"), c(5, 5, 5, 5))
  )
contr_bw <- calc_contrast(variable = Val,
                 between = Let,
                 between_levels = c("A", "B", "C", "D"),
                 lambda_between = c(-3, -1, 1, 3),
                 data = tab31)
contr_bw
summary(contr_bw)

# Exampel for within-subjects-design Calculation 16.6 from
# Sedlmeier and Renkewitz (2018, p. 537)

sedlmeier537 <- data.frame(
   Var = c(27, 25, 30, 29, 30, 33, 31, 35,
            25, 26, 32, 29, 28, 30, 32, 34,
            21, 25, 23, 26, 27, 26, 29, 31,
            23, 24, 24, 28, 24, 26, 27, 32),
   within = as.factor(rep(1:4,c(8,8,8,8))),
   ID = as.factor(rep(1:8,4))
 )
contr_wi <- calc_contrast(
   variable = Var,
   within = within,
   within_levels = as.character(c(2,3,1,4)),
   ID = ID,
   lambda_within = c(0.25,-.75,1.25,-.75),
   data=sedlmeier537
 )
contr_wi
summary(conr_wi, ci=.90)

# Exampel for mixed-designs Table 5.3 from
# Rosenthal, Rosnow and Rubin (2001)

}
\references{
Rosenthal, R., Rosnow, R.L., & Rubin, D.B. (2000).
Contrasts and effect sizes in behavioral research:
A correlational approach. New York: Cambridge University Press.
}
