% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_contrast_aggregated.R
\name{calc_contrast_aggregated}
\alias{calc_contrast_aggregated}
\title{Calculate between contrast analysis from aggregated data (means, sds and ns)}
\usage{
calc_contrast_aggregated(means, sds, ns, lambdas, between, data)
}
\arguments{
\item{means}{named character, vector of mean values for every condition in
data}

\item{sds}{named character, vector of standard deviation values for every
condition in data}

\item{ns}{named character, vector of sample size values for every condition
in data}

\item{lambdas}{contrast weights must be a named numeric.
Names must match the levels of \code{between}. If
\code{lambda_between}
does not sum up to zero, this will be done automatically.}

\item{between}{independent variable that divides the data into
independent groups (a factor).}

\item{data}{optional argument for the \code{data.frame} containing all variables
except for lambdas}
}
\value{
Calculates the significance of the contrast analysis. The
}
\description{
Calculate between contrast analysis from aggregated data (means, sds and ns)
}
\examples{

}
\references{
Rosenthal, R., Rosnow, R.L., & Rubin, D.B. (2000). Contrasts and
  effect sizes in behmeansioral research: A correlational approach. New York:
  Cambridge University Press.
}
