% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cofad-package.R
\docType{package}
\name{cofad-package}
\alias{cofad}
\alias{cofad-package}
\title{cofad: Contrast Analyses for Factorial Designs}
\description{
Contrast analysis for factorial designs is an alternative to the classical ANOVA approach with the advantage of testing focused hypothesis. The method is mainly based on Rosenthal, Rosnow and Rubin (2000, ISBN:978-0521659802) and Sedlmeier and Renkewitz (2018, ISBN:978-3868943214).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/johannes-titz/cofad}
}

}
\author{
\strong{Maintainer}: Markus Burkhardt \email{markus.burkhardt@psychologie.tu-chemnitz.de}

Authors:
\itemize{
  \item Johannes Titz \email{johannes.titz@gmail.com}
}

}
\keyword{internal}
