% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.cofad_mx}
\alias{print.cofad_mx}
\title{Output of a mixed design contrast analysis}
\usage{
\method{print}{cofad_mx}(x, ...)
}
\arguments{
\item{x}{output of calc_contrast}

\item{...}{further arguments}
}
\value{
Displays the significance of the contrast analysis. The
contrastweights, the corresponding group and an effectsize are given.
}
\description{
Output of a mixed design contrast analysis
}
