% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rosenthal_tbl59}
\alias{rosenthal_tbl59}
\title{Children data set by Rosenthal and Rosnow (2000)}
\format{
a data frame with 36 rows and 4 columns
\describe{
  \item{dv}{dependent variable}
  \item{between}{age group (8, 10, 12 years)}
  \item{ID}{unique identifier for child}
  \item{within}{measurement (1, 2, 3, 4)}
}
}
\source{
Rosenthal, R., Rosnow, R. L., & Rubin, D. B. (2000). Contrasts and
  Effect Sizes in Behavioral Research: A Correlational Approach. Cambridge
  University Press.
}
\usage{
rosenthal_tbl59
}
\description{
Table 5.3 in Rosenthal and Rosnow (2000) on p. 129.
}
\keyword{datasets}
