% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sedlmeier_p525}
\alias{sedlmeier_p525}
\title{Problem solving data set by Sedlmeier & Renkewitz (2018)}
\format{
a data frame with 15 rows and 3 columns
\describe{
  \item{lsg}{dv, number of svoled exercises}
  \item{between}{group, KT=no training, JT=boys-specific, MT=girls-specific}
  \item{lambda}{lambdas used for this example}
}
}
\source{
Sedlmeier, P., & Renkewitz, F. (2018). Forschungsmethoden und
  Statistik für Psychologen und Sozialwissenschaftler (3rd ed.). Pearson
  Studium.
}
\usage{
sedlmeier_p525
}
\description{
Example 16.2, table 16.1 in Sedlmeier & Renkewitz (2018). Fictious data set
with 15 boys divided into three groups (no training, boys-specific material,
girls-specific traning material). The DV is the number of solved problem
(similar to the training).
}
\keyword{datasets}
