% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary.cofad_bw}
\alias{summary.cofad_bw}
\title{Summary of between subject design contrast analysis}
\usage{
\method{summary}{cofad_bw}(object, type = "Contrast Analysis Between", ...)
}
\arguments{
\item{object}{output of calc_contrast}

\item{...}{further arguments}
}
\value{
Displays type of contrast analysis, lambdas, t-table, ANOVA table and
  typical effect sizes. If you assign this to a variable, it will be a list
  with the elements Lambdas, tTable, FTable, Effects.
}
\description{
Summary of between subject design contrast analysis
}
