% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary.cofad_wi}
\alias{summary.cofad_wi}
\title{Summary of within subject design contrast analysis}
\usage{
\method{summary}{cofad_wi}(object, ci = 0.95)
}
\arguments{
\item{object}{output of calc_contrast}

\item{ci}{confidence intervall for composite Score (L-Values)}
}
\value{
Displays ANOVA table of the contrastanalysis
and the typical effectsizes.
}
\description{
Summary of within subject design contrast analysis
}
